Remove duplicate matrix multiplication code.
The first call to `Matrix.multiply` forces the initialization of the
various default matrices. Their initialization code may call a
deprecated function, which in turn may cause a runtime exception. This
was the reason for the duplicate code in the first place. Making them
`lazy` seems to solve this problem.